+++ core/filter.h
@@ -26,6 +26,7 @@ namespace oidn {
     virtual void setImage(const std::string& name, const Image& data) = 0;
     virtual void removeImage(const std::string& name) = 0;
     virtual void setData(const std::string& name, const Data& data) = 0;
+    virtual void updateData(const std::string& name) = 0;
     virtual void removeData(const std::string& name) = 0;
     virtual void set1i(const std::string& name, int value) = 0;
     virtual int get1i(const std::string& name) = 0;